requirements.txt
annotated-types==0.6.0
attrs==23.1.0
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
et-xmlfile==1.1.0
filelock==3.13.1
greenlet==3.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.2
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
parsel==1.8.1
pdf2image==1.16.3
Pillow==10.1.0
pluggy==1.3.0
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==23.3.0
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.3
pytest-dotenv==0.5.2
pytest-ordering==0.6
python-dateutil==2.8.2
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
service-identity==23.1.0
setuptools==69.0.2
six==1.16.0
SQLAlchemy==2.0.23
tldextract==5.1.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.8.0
urllib3==2.1.0
w3lib==2.1.2
zope.interface==6.1
